+++ src/gufo/thor/services/bi.py
@@ -11,15 +11,21 @@ Attributes:
 """
 
 # Gufo Thor modules
+from .auth import auth
 from .clickhouse import clickhouse
+from .envoy import envoy
+from .login import login
+from .migrate import migrate
 from .noc import NocService
+from .static import static
 
 
 class BiService(NocService):
     """bi service."""
 
     name = "bi"
-    dependencies = (clickhouse,)
+    dependencies = (auth, clickhouse, envoy, login, migrate, static)
+    expose_http_prefix = "/api/bi/"
 
 
 bi = BiService()