+++ src/gufo/thor/services/envoy.py
@@ -1,13 +1,13 @@
 # ---------------------------------------------------------------------
-# Gufo Thor: nginx service
+# Gufo Thor: envoy service
 # ---------------------------------------------------------------------
 # Copyright (C) 2023, Gufo Labs
 # ---------------------------------------------------------------------
 """
-nginx service.
+envoy service.
 
 Attributes:
-    nginx: nginx service singleton.
+    eenvoy: envoy service singleton.
 """
 
 # Python modules
@@ -23,10 +23,9 @@ from ..error import CancelExecution
 from ..log import logger
 from ..utils import write_file
 from .base import BaseService
-from .login import login
-from .traefik import traefik
 
 HTTP_OK = 200
+HTTPS = 443
 
 
 @dataclass
@@ -54,61 +53,129 @@ DOMAINS = {
 }
 
 
-class NginxService(BaseService):
-    """nginx service."""
+@dataclass
+class Route(object):
+    """
+    routes part of config.
 
-    name = "nginx"
-    dependencies = (traefik, login)
-    compose_image = "nginx:stable"
-    compose_volumes = ["./etc/nginx:/etc/nginx"]
-    SUBJ_PATH = "etc/nginx/ssl/domain_name.txt"
+    Attributes:
+        name: Service name.
+        prefix: HTTP prefix.
+        disable_auth: Disable external authorization.
+        prefix_rewite: Rewrite prefix, if set.
+        redirect_to: Redirect to path, if matched.
+    """
+
+    name: str
+    prefix: str
+    disable_auth: bool
+    prefix_rewrite: Optional[str] = None
+    redirect_to: Optional[str] = None
+
+
+class EnvoyService(BaseService):
+    """envoy service."""
+
+    name = "envoy"
+    compose_image = "envoyproxy/envoy:v1.28.0"
+    # compose_depends_condition = ComposeDependsCondition.HEALTHY
+    # compose_healthcheck = {
+    #     "test": ["CMD", "envoy", "healthcheck", "--ping"],
+    #     "interval": "3s",
+    #     "timeout": "2s",
+    #     "retries": 3,
+    # }
+    compose_volumes = ["./etc/envoy/:/etc/envoy/:ro"]
+    SUBJ_PATH = Path("etc", "envoy", "ssl", "domain_name.txt")
     RSA_KEY_SIZE = 4096
     CERT_DAYS = 3650
 
     def get_compose_networks(
-        self: "NginxService", config: Config, svc: Optional[ServiceConfig]
+        self: "EnvoyService", config: Config, svc: Optional[ServiceConfig]
     ) -> Optional[Dict[str, Any]]:
         """Get networks section."""
         return {
-            "aliases": ["nginx", config.expose.domain_name],
+            "aliases": ["envoy", config.expose.domain_name],
         }
 
     def get_compose_ports(
-        self: "NginxService", config: Config, svc: Optional[ServiceConfig]
+        self: "EnvoyService", config: Config, svc: Optional[ServiceConfig]
     ) -> Optional[List[str]]:
         """Get ports section."""
         return [f"{config.expose.port}:443"]
 
     def prepare_compose_config(
-        self: "NginxService", config: Config, svc: Optional[ServiceConfig]
+        self: "EnvoyService",
+        config: Config,
+        svc: Optional[ServiceConfig],
+        services: List["BaseService"],
     ) -> None:
-        """Geerate configuration files."""
-        cfg_root = Path("etc", "nginx")
-        self.render_file(cfg_root / "nginx.conf", "nginx.conf")
-        self.render_file(cfg_root / "mime.types", "mime.types")
+        """Generate config."""
+        # Generate domain_name_and_port
+        if config.expose.port == HTTPS:
+            domain_name_and_port = config.expose.domain_name
+        else:
+            domain_name_and_port = (
+                f"{config.expose.domain_name}:{config.expose.port}"
+            )
+        # Generate routes
+        routes: List[Route] = []
+        for s in services:
+            prefix = s.get_expose_http_prefix(config, None)
+            if not prefix:
+                continue
+            routes.append(
+                Route(
+                    name=s.name,
+                    prefix=prefix,
+                    disable_auth=not s.require_http_auth,
+                    prefix_rewrite=s.rewrite_http_prefix,
+                )
+            )
+            if s.name == "web":
+                # Add redirect to desktop
+                routes.append(
+                    Route(
+                        name="Redirect / -> /main/desktop/",
+                        prefix="/",
+                        disable_auth=True,
+                        redirect_to="/main/desktop/",
+                    )
+                )
+        routes = sorted(
+            routes,
+            key=lambda x: "---".join(
+                y if y else "zzzzzzzz" for y in x.prefix.split("/")
+            )
+            + f"---{not bool(x.redirect_to)}",
+        )
+        # Add desktop redirect, if necessary
+        #
         self.render_file(
-            cfg_root / "conf.d" / "noc.conf",
-            "noc.conf",
+            Path("etc", "envoy", "envoy.yaml"),
+            "envoy.yaml",
             domain_name=config.expose.domain_name,
-            port=config.expose.port,
+            domain_name_and_port=domain_name_and_port,
+            routes=routes,
+            services=sorted({r.name for r in routes if not r.redirect_to}),
         )
         # Prepare TLS certificates
         if self._to_rebuild_certificate(config):
             self._rebuild_certificate(config)
 
-    def _to_rebuild_certificate(self: "NginxService", config: Config) -> bool:
+    def _to_rebuild_certificate(self: "EnvoyService", config: Config) -> bool:
         """Check if SSL certificate must be rebuilt."""
         if not os.path.exists(self.SUBJ_PATH):
             return True
         with open(self.SUBJ_PATH) as fp:
             return fp.read() != self.get_cert_subj(config)
 
-    def get_cert_subj(self: "NginxService", config: Config) -> str:
+    def get_cert_subj(self: "EnvoyService", config: Config) -> str:
         """Get certificate subj."""
         return f"CN={config.expose.domain_name}"
 
     def _get_signed_csr(
-        self: "NginxService", csr_proxy: str, csr: bytes
+        self: "EnvoyService", csr_proxy: str, csr: bytes
     ) -> bytes:
         """
         Sign certificate via CSR Proxy.
@@ -129,13 +196,13 @@ class NginxService(BaseService):
         logger.error("Failed to sign certificate")
         raise CancelExecution()
 
-    def _rebuild_certificate(self: "NginxService", config: Config) -> None:
+    def _rebuild_certificate(self: "EnvoyService", config: Config) -> None:
         """Rebuild SSL certificates."""
         from gufo.acme.clients.base import AcmeClient
 
-        key_path = Path("etc", "nginx", "ssl", "noc.key")
-        csr_path = Path("etc", "nginx", "ssl", "noc.csr")
-        cert_path = Path("etc", "nginx", "ssl", "noc.crt")
+        key_path = Path("etc", "envoy", "ssl", "noc.key")
+        csr_path = Path("etc", "envoy", "ssl", "noc.csr")
+        cert_path = Path("etc", "envoy", "ssl", "noc.crt")
 
         di = DOMAINS.get(config.expose.domain_name)
         if di is None:
@@ -169,4 +236,4 @@ class NginxService(BaseService):
             fp.write(self.get_cert_subj(config))
 
 
-nginx = NginxService()
+envoy = EnvoyService()