+++ src/gufo/thor/services/shell.py
@@ -27,7 +27,7 @@ class ShellService(NocService):
     """web service."""
 
     name = "shell"
-    dependencies = (migrate, postgres, mongo, clickhouse, worker)
+    dependencies = (clickhouse, migrate, mongo, postgres, worker)
     compose_extra = {"scale": 0}
 
     def get_compose_command(