+++ src/gufo/thor/services/topo.py
@@ -13,6 +13,7 @@ Attributes:
 # Gufo Thor modules
 from .datastream import datastream
 from .liftbridge import liftbridge
+from .migrate import migrate
 from .noc import NocService
 
 
@@ -20,7 +21,7 @@ class TopoService(NocService):
     """topo service."""
 
     name = "topo"
-    dependencies = (datastream, liftbridge)
+    dependencies = (datastream, liftbridge, migrate)
 
 
 topo = TopoService()